Refactor level logic
