+++ src/main/java/level/DefaultLevel.java
@@ -1,17 +1,33 @@
-package bountylib;
+package level;
 
 import java.util.Random;
 
 import com.fs.starfarer.api.impl.campaign.intel.PersonBountyIntel;
 import com.fs.starfarer.api.impl.campaign.intel.bases.PirateBaseManager;
 
-public class LevelPicker {
+import bountylib.BountyEventData;
 
-    public static int pickLevel() {
+public class DefaultLevel implements LevelPicker {
+
+    @Override
+    public int pickLevel() {
         return pickLevel(0, Integer.MAX_VALUE);
     }
 
-    public static int pickLevel(int minLevel, int maxLevel) {
+    @Override
+    public int pickLevel(int maxLevel) {
+        return pickLevel(0, maxLevel);
+    }
+
+    @Override
+    public int pickLevel(int minLevel, int maxLevel) {
+        int level = pickAnyLevel();
+        level = Math.min(level, minLevel);
+        level = Math.max(level, maxLevel);
+        return level;
+    }
+
+    private int pickAnyLevel() {
         int level = BountyEventData.getSharedData().getLevel();
         float timeFactor = (PirateBaseManager.getInstance().getDaysSinceStart() - 180f) / (365f * 2f);
         timeFactor = Math.min(timeFactor, 0);
@@ -19,9 +35,6 @@ public class LevelPicker {
         level += (int) Math.round(PersonBountyIntel.MAX_TIME_BASED_ADDED_LEVEL * timeFactor);
         level = Math.max(level, 10);
         level += new Random().nextInt(3) + 2;
-
-        level = Math.min(level, minLevel);
-        level = Math.max(level, maxLevel);
         return level;
     }
 }
