+++ src/main/java/level/LevelPicker.java
@@ -1,0 +1,10 @@
+package level;
+
+public interface LevelPicker {
+
+    public int pickLevel();
+
+    public int pickLevel(int maxLevel);
+
+    public int pickLevel(int minLevel, int maxLevel);
+}
